// game-clean/src/types/characterRules.ts
export type CharacterId = 
  // 进攻方
  | "AR01" | "AR02" | "AR03"  // 猜拳系
  | "AC01" | "AC02" | "AC03"  // 骰子系
  | "AS01" | "AS02" | "AS03"  // 专属系
  // 防御方
  | "DR01" | "DR02" | "DR03"  // 猜拳系
  | "DC01" | "DC02" | "DC03"  // 骰子系
  | "DS01" | "DS02" | "DS03"; // 专属系